requirements.txt
fastparquet
matplotlib
numpy
pandas
pyarrow
python-dateutil
pytorch-ignite
tensorboard
torch
torchvision
tqdm